requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
torch==2.0.1+cpu --extra-index-url https://download.pytorch.org/whl/cpu
torch-geometric==2.3.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
transformers==4.33.2
sentencepiece==0.1.99
biopython==1.81
requests==2.31.0